app/authService/requirements.txt
fastapi==0.136.0
uvicorn==0.35.0
pydantic==2.13.2
email-validator==2.3.0
python-dotenv==1.2.2
psycopg[binary]==3.2.10
bcrypt==5.0.0
PyJWT==2.12.1
